Finish exercise #3, similar to solution
At first I didn't really group the two elements, just createad
a third class and added to both elements in the HTML text.
It delivered me the correct result, although after checking
the solution, I realized that the whole purpose of the exercise
was to do it using only two classes. Note to self: read the
instructions with more attention next time.