Finish CSS Flex exercise 4 with notes
Managed to get to the desired result but used more coding than needed.
Instead of justifying flex content to center, I was trying to use auto
margins on the sides; also, if I had used text-align: center; on body
I would not have to use it twice on title and descriptive texts.
Other than that, achieved the desired outcome. On we go.