Finish CSS Flex exercise 5 correctly
This one had me on a "it all makes sense now" moment
when I realized that I did not have to use flex for the container
with the the header, text and buttons. Took me quite a while to figure
it out, especially since we're dealing with Flex so much right now.
On we go!